requirements.txt
streamlit==1.38.0
openai==1.50.2
python-dotenv==1.0.1
nltk==3.9.1
pydantic==2.9.2
email-validator==2.2.0
phonenumbers==8.13.47
tqdm==4.66.5
